feature: 한글 하드 코딩 수정, 인터넷 퍼미션 제거 #83
