+++ app/src/main/AndroidManifest.xml
@@ -2,8 +2,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.kdjj.ratatouille">
 
-    <uses-permission android:name="android.permission.INTERNET" />
-
     <application
         android:name=".app.App"
         android:allowBackup="true"
