+++ presentation/src/main/res/layout/item_others_recipe.xml
@@ -26,7 +26,7 @@
                 android:id="@+id/imageView_others_item_img"
                 android:layout_width="120dp"
                 android:layout_height="120dp"
-                android:scaleType="fitXY"
+                android:scaleType="centerCrop"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toTopOf="parent"
